ci: removed stress test
The stress test was supposed to test the performance of the different server implementations of connaisseur (cheroot vs uwsgi vs ...). Since we settled with cheroot for now, the stress test isn't needed anymore for the regular pipeline run, especially because it's result was discarded anyways. In order to have a "successful" pipeline again, the stress test was removed fromthe regular dev run and instead moved to the nightly run, so we can still observe stability issues.